gaitmap-challenges/pyproject.toml
[tool.poetry]
name = "gaitmap-challenges"
version = "0.1.0"
description = "A set of benchmark challenges for IMU based human gait analysis"
authors = ["Arne Küderle <arne.kuederle@fau.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pandas = ">=1.4.3"
scikit-learn = ">=1.0"
gaitmap = {git = "https://github.com/mad-lab-fau/gaitmap.git", rev = "master", extras = ["stats"]}
optuna = ">=3.0.0"
gaitmap-datasets = ">=0.5.0"
holoviews = {version = "^1.15.3", optional = true}
py-cpuinfo = "^9.0.0"

[tool.poetry.extras]
visualization = ["holoviews"]

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
poethepoet = "^0.16.5"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
prospector = "^1.8.3"
ipykernel = "^6.19.2"
numpydoc = "^1.5.0"
Sphinx = "^5.3.0"
sphinx-gallery = "^0.11.1"
recommonmark = "^0.7.1"
memory-profiler = "^0.61.0"
matplotlib = "^3.6.2"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | docs
    | build
    | dist
    | \.virtual_documents
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 120
skip_gitignore = true


[tool.poe.tasks]
_format_black = "black ."
_format_isort = "isort ."
format = { sequence = ["_format_black", "_format_isort"], help = "Format all files." }
lint = { cmd = "prospector --no-autodetect", help = "Lint all files with Prospector." }
_check_black = "black . --check"
_check_isort = "isort . --check"
check = { sequence = ["_check_black", "_check_isort", "lint"], help = "Check all potential format and linting issues." }
test = { cmd = "pytest --cov=gaitmap-challenges --cov-report=term-missing --cov-report=xml", help = "Run Pytest with coverage." }
version = { script = "_tasks:task_update_version()" }
bump_dev = { script = "_tasks:task_bump_all_dev()" }
